refactor project filters, add filter reset
