+++ web-components/src/components/organisms/ProjectFilters/ProjectFilters.stories.tsx
@@ -35,6 +35,7 @@ const Template: StoryFn = args => {
       filters={args.filters}
       activeFilterIds={activeFilters}
       onFilterChange={onFilterChange}
+      onFilterReset={() => setActiveFilters([])}
     />
   );
 };
